feat(build): update build configuration for ADOC project
This commit updates the build configuration to reflect the new ADOC
project. Changes include modifying artifact and Algolia details in
the GitHub Actions workflow and adding a new build profile in
`buildprofiles.xml` for the ADOC instance. These updates aim to
ensure the correct deployment and indexing of the ADOC documentation
with Algolia search integration.